This is a zip of all the files I extracted from the bios.
This is all the the modules, Dxe, and images I managed to extract from the stock bios file after decompressing it. Just a break down of me tearing the bios apart and ready getting in there to go through all the settings and options it has to offer and make it as solid as possible so I have issues with it long term. (Contains .efi, .fbd, .bin, i.ifr, .csv files that I had to extract and patch.)